Refactor ask_availability and add mini app url
