+++ bot/requirements.txt
@@ -1,4 +1,4 @@
-pyTelegramBotAPI>=4.14.0
+pyTelegramBotAPI
 python-dotenv
 requests>=2.31.0
 urllib3>=2.1.0
