+++ bot/telegram/handlers/availability_handlers.py
@@ -8,8 +8,12 @@ from ..services.database_service import getEntry, setEntry, updateEntry
 from ..services.scheduling_service import format_availability_summary
 from ..utils.web_app import create_web_app_url
 from ..services.event_service import getEvent, getUserAvailability, updateUserAvailability
+
+# Import from services
+from ..services.availability_service import ask_availability
+
+# Import from utils
 from ..utils.message_templates import HELP_MESSAGE
-#from ..utils.native_interface import create_native_availability_selector, handle_native_availability_callback
 
 logger = logging.getLogger(__name__)
 
@@ -75,36 +79,3 @@ def register_availability_handlers(bot):
             bot.reply_to(message, "Failed to update availability. Please try again later.")
 
     return bot
-
-def ask_availability(chat_id: int, event_id: str, username: str = None):
-    """Ask user to provide availability for an event"""
-    try:
-        # Get event details
-        event = getEvent(event_id)
-        if not event:
-            bot.send_message(chat_id, "Event not found")
-            return
-        
-        # Create web app button for availability selection
-        if username:
-            web_app_url = create_web_app_url("/dragselector", 1, event_id=event_id, username=username)
-        else:
-            # Fallback for when username is not provided
-            web_app_url = create_web_app_url("/dragselector", 1, event_id=event_id)
-        
-        markup = types.InlineKeyboardMarkup()
-        webapp_btn = types.InlineKeyboardButton(
-            text="Select Availability",
-            web_app=types.WebAppInfo(url=web_app_url)
-        )
-        markup.add(webapp_btn)
-        
-        bot.send_message(
-            chat_id,
-            f"Please select your availability for {event['event_name']}:",
-            reply_markup=markup
-        )
-        
-    except Exception as e:
-        logger.error(f"Error in ask_availability: {str(e)}")
-        bot.send_message(chat_id, "Failed to set up availability selection. Please try again later.")
