+++ bot/telegram/handlers/event_handlers.py
@@ -12,7 +12,8 @@ from ..services.user_service import updateUsername
 from ..services.event_service import getEventSleepPreferences, getUserAvailability, updateUserAvailability, create_event, get_event_by_id, join_event
 from ..services.scheduling_service import calculate_optimal_meeting_time
 from ..utils.date_utils import daterange
-from .availability_handlers import ask_availability
+
+from ..services.availability_service import ask_availability
 
 # Keep track of processed message IDs to prevent duplicate processing
 processed_messages = set()
