+++ bot/telegram/handlers/group_handlers.py
@@ -5,7 +5,7 @@ import re
 
 # Import from config and services
 from ..config.config import bot
-from .availability_handlers import ask_availability
+from ..services.availability_service import ask_availability
 
 def register_group_handlers(bot):
     """Register all group-related handlers"""
@@ -17,7 +17,7 @@ def register_group_handlers(bot):
         if match:
             event_id = match.group(1)
             print(event_id)
-            ask_availability(message.chat.id, event_id)
+            ask_availability(message.chat.id, event_id, group=True)
         else:
             bot.reply_to(message, "Invalid event ID.")
